Extract calls from Client.main() to a ClientRunner
